# backend/swagger-ui/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Baustellenverwaltung",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/task": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "Returns all task for all construction sites",
        "description": "Returns a array of all tasks",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "Task"
      }
    },
    "/user/{userId}/task": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Returns all task for user",
        "description": "Returns a array of all tasks",
        "operationId": "getTasksByuserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "Task"
      }
    },
...
    "/constructionSite/{constructionSiteId}/task": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "Returns all tasks for specified construction sites",
        "description": "Returns a array of tasks",
        "operationId": "getConstructionSiteTasks",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "Task"
      },
      "post": {
        "tags": [
          "task"
        ],
        "summary": "Add task to constructionSite",
        "operationId": "addTask",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Task object that should be added to constructionSite",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid properties on Task",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Task"
      }
    },
    "/constructionSite/{constructionSiteId}/task/{taskId}": {
      "put": {
        "tags": [
          "task"
        ],
        "summary": "Update task by ID",
        "description": "Update task",
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "ID of task that should be updated",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Task object that should be added to constructionSite",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Task not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Task"
      },
      "delete": {
        "tags": [
          "task"
        ],
        "summary": "Delete task by ID",
        "description": "Deletes Task from construction Site",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "ID of task that needs to be deleted",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {}
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Task not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "Task"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Task": {
        "required": [
          "assignedTo",
          "priority",
          "status",
          "title",
          "constructionSiteId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Task_createdBy"
          },
          "modifiedAt": {
            "type": "string",
            "example": "{}"
          },
          "createdAt": {
            "type": "string",
            "example": "{}"
          },
          "constructionSiteId": {
            "type": "number",
            "example": 0
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "normal",
              "low"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "completed"
            ]
          },
          "assignedTo": {
            "$ref": "#/components/schemas/Task_createdBy"
          }
        },
        "example": {
          "createdAt": "{}",
          "title": "title",
          "createdBy": {
            "user": "name",
            "userId": "userId"
          },
          "modifiedAt": "{}",
          "id": 0,
          "text": "text",
          "priority": "high",
          "assignedTo": "{}",
          "status": "open"
        },
        "xml": {
          "name": "Task"
        }
      },
...
      "Task_createdBy": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "name of user"
          },
          "userId": {
            "type": "string",
            "description": "userId reference"
          }
        },
        "example": {
          "name": "name",
          "userId": "userId"
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}